dsktool: diskapi - modify fdc_read_n_bytes to allow reading more than 1 sector on the same call (WIP)
When reading full sectors, the memory after the last byte (we provide the
size in DE) may be overwritten up to 511 bytes.  Modify fdc_read_n_bytes to
use sector read buffer for saving that area before reading a complete
sector, and restore it afterwards.

This will allow this function to read more than 1 sector, instead of using
this function to read the track 1 sector at a time.